+++ tests/modules/e2e_runtime.lua
@@ -41,6 +41,23 @@ function print_r(arr, indentLevel)
   return str
 end
 
+-- qwertyuiopasdfghjklzxcvbnm
+for i = 97, 122 do table.insert(charset, string.char(i)) end
+
+function string.random(length)
+  math.randomseed(os.time())
+
+  if length > 0 then
+    return string.random(length - 1) .. charset[math.random(1, #charset)]
+  else
+    return ""
+  end
+end
+
+function string.ends(str, with)
+  return with == '' or string.sub(str, -string.len(with)) == with
+end
+
 --[[
   Nakama module
 ]]--
@@ -324,4 +341,89 @@ do
   local objectDecode = nk.base16_decode(objectEncode)
   assert(objectDecode, "'objectDecode' must not be nil")
   assert(objectDecode == '{"hello": "world"}', '"objectDecode" must equal {"hello": "world"}')
+end
+
+-- sql_exec and sql_query
+do
+  -- Table names cannot start with a number so we can't use our usual UUID here.
+  local t = string.random(20)
+
+  local query = "CREATE TABLE " .. t .. " ( foo VARCHAR(20), bar BIGINT )"
+  local params = {}
+  local status, result = pcall(nk.sql_exec, query, params)
+  if not status then
+    print(result)
+  end
+  assert(result == 0)
+
+  local query = "INSERT INTO " .. t .. " (foo, bar) VALUES ($1, $2), ($3, $4), ($5, $6)"
+  local params = {"foo1", 1, "foo2", 2, "foo3", 3}
+  local status, result = pcall(nk.sql_exec, query, params)
+  if not status then
+    print(result)
+  end
+  assert(result == 3)
+
+  local query = "SELECT * FROM " .. t .. " WHERE bar = $1"
+  local params = {2}
+  local status, result = pcall(nk.sql_query, query, params)
+  if not status then
+    print(result)
+  end
+  assert(#result == 1)
+  assert(result[1].foo == "foo2")
+  assert(result[1].bar == 2)
+
+  local query = "SELECT * FROM " .. t .. " WHERE bar >= $1 ORDER BY bar DESC"
+  local params = {2}
+  local status, result = pcall(nk.sql_query, query, params)
+  if not status then
+    print(result)
+  end
+  assert(#result == 2)
+  assert(result[1].foo == "foo3")
+  assert(result[1].bar == 3)
+  assert(result[2].foo == "foo2")
+  assert(result[2].bar == 2)
+
+  local query = "DELETE FROM " .. t .. " WHERE bar = $1"
+  local params = {2}
+  local status, result = pcall(nk.sql_exec, query, params)
+  if not status then
+    print(result)
+  end
+  assert(result == 1)
+
+  local status, result = pcall(nk.sql_exec, query, params)
+  if not status then
+    print(result)
+  end
+  assert(result == 0)
+
+  local query = "SELECT * FROM " .. t .. " WHERE bar >= $1 ORDER BY bar DESC"
+  local params = {2}
+  local status, result = pcall(nk.sql_query, query, params)
+  if not status then
+    print(result)
+  end
+  assert(#result == 1)
+  assert(result[1].foo == "foo3")
+  assert(result[1].bar == 3)
+
+  local query = "DROP TABLE " .. t
+  local params = {}
+  local status, result = pcall(nk.sql_exec, query, params)
+  if not status then
+    print(result)
+  end
+  assert(result == 0)
+
+  local query = "SELECT * FROM " .. t
+  local params = {}
+  local status, result = pcall(nk.sql_query, query, params)
+  if not status then
+    print(result)
+  end
+  assert(not status)
+  assert(string.ends(result, 'sql query error: pq: table "' .. t .. '" does not exist'))
 end