+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [keep a changelog](http://keepachangelog.com/) and this p
 ### Added
 - New code runtime function to list leaderboard records for a given set of users.
 - New code runtime function to list leaderboard records around a given user.
+- New code runtime function to execute raw SQL queries. 
 
 ### Changed
 - Handle update now returns a bad input error code if handle is too long.